Revert "Prepare upb_value for encoding/decoding map."
This reverts commit f30dd0ff0cead9bb1f8f3b1b295392d75705486b.
